[Android] In Progress VoteDetailScreen.

+++ android/core/data/vote-contract/src/main/java/dylan/kwon/votechain/core/data/vote_contract/model/Voter.kt
@@ -1,9 +1,7 @@
 package dylan.kwon.votechain.core.data.vote_contract.model
 
-import java.math.BigInteger
-
 
 data class Voter(
     val id: String,
-    val votings: List<BigInteger>
+    val votings: List<Long>
 )
